actualizacion

+++ Vagrant.bootstrap.sh
@@ -68,7 +68,9 @@ if [ ! -x "$(command -v docker)" ] ; then
 	sudo apt-get update -y
 
 	#Instalo docker desde el repositorio oficial
-        sudo apt-get install -y docker-ce docker-compose
+        sudo apt-get install -y docker-ce docker-compose golang-github-docker-libnetwork-dev \
+		golang-github-containerd-docker-containerd-dev golang-github-docker-engine-api-dev ruby-docker-api \
+		docker-registry libnss-docker
 
         #Lo configuro para que inicie en el arranque
         sudo systemctl enable docker
@@ -85,28 +87,50 @@ pwd
 echo " "
 ls -ltra
 
-sudo docker ps -a | grep -v CONTAINER | while read line ; do sudo docker rm `echo $line | awk '{print$1}'` ; done
-sudo docker images | grep -v REPOSITORY | while read line ; do sudo docker image rm  `echo $line | awk '{print$3}'` ; done
+echo " ========================================================================" 
+echo " "
+echo " reconstruyendo docker"
+if [ `sudo docker ps | wc -l` -gt 1 ] ; then  
+	sudo docker ps | grep -v CONTAINER | awk '{print$1}' | while read line ; do sudo docker stop $line ; done
+	sudo docker ps -a | grep -v CONTAINER | while read line ; do sudo docker rm `echo $line | awk '{print$1}'` ; done
+	sudo docker images | grep -v REPOSITORY | while read line ; do sudo docker image rm  `echo $line | awk '{print$3}'` ; done
+fi
+
+
+echo " ========================================================================"
 
 echo " "
-echo "creando docker"
-sudo docker-compose up -d
+echo "creando docker DB"
+sudo docker-compose stop && docker-compose rm && docker-compose build && docker-compose up -d 	
 
 echo " "
-echo "docker creados"
+echo "docker activos "
 sudo docker ps
+echo "docker creados "
 sudo docker ps -a
 
-echo " ------------------------------------------------------------- "
+echo " ========================================================================"
 echo "IP DE BASE DE DATOS"
 sudo docker inspect `sudo docker ps | grep mysql | awk '{print$1}'` | grep IPAddress | tail -1 | awk '{print$2}' | sed 's/\"//g' | sed 's/\,//g'
-
-echo " ------------------------------------------------------------- "
+IP_BD=`sudo docker inspect \`sudo docker ps | grep mysql | awk '{print$1}'\` | grep IPAddress | tail -1 | awk '{print$2}' | sed 's/\"//g' | sed 's/\,//g'`
+echo " ========================================================================"
 echo "IP WEB SERVER"
 sudo docker inspect `sudo docker ps | grep php | awk '{print$1}'` | grep IPAddress | tail -1 | awk '{print$2}' | sed 's/\"//g' | sed 's/\,//g'
 
+# echo " ========================================================================"
+# echo "Adecuando db_connect a la IP del contenedor de BD"
+# sudo sed -i 's/127.0.0.1/'${IP_DB}'/' /var/www/myapp/src/include/db_connect.php
+# #echo " "
+# #sudo docker exec -i apache2_php cd /var/www/html/myapp
 
-#echo " "
-#sudo docker exec -i apache2_php cd /var/www/html/myapp
+echo " ========================================================================"
 
-sudo docker exec -i db_mysql mysql -uroot -proot devops_app < /vagrant/docker/configs/mysql/script.sql
+echo "creando DB"
+if [ `sudo su - ; ls /var/db/mysql/devops_app | grep welcome | wc -l | awk '{print$1}'` -gt 2  ] ; then
+	echo "db y tabla creada"
+else
+	sudo docker exec -i dbmysql mysql -uroot -proot devops_app < /vagrant/docker/configs/mysql/script.sql
+fi
+
+
+
